fixed bug in printf added printing numbers in base 10 and 16

+++ libc/stdio/puts.cpp
@@ -1,5 +1,14 @@
 #include <stdio.h>
  
-int puts(const char* string) {
+/**
+ * Puts a string to stdout.
+ *
+ * Writes the string pointed to by string to stdout, followed by a newline character.
+ *
+ * @param string The string to write.
+ * @return The number of characters written if successful, or EOF on error.
+ */
+size_t puts(const char *string)
+{
 	return printf("%s\n", string);
 }
